requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
numpy>=1.21.0
psycopg2>=2.9
sqlalchemy>=2.0
